+++ src/main/java/gift/web/exception/GlobalExceptionHandler.java
@@ -1,5 +1,9 @@
 package gift.web.exception;
 
+import gift.web.exception.forbidden.ForbiddenException;
+import gift.web.exception.kakaoapi.KakaoApiException;
+import gift.web.exception.notfound.NotFoundException;
+import gift.web.exception.unauthorized.UnauthorizedException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
@@ -9,34 +13,17 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ResponseEntity<?> handleProductNotFoundException(ProductNotFoundException e) {
+
+    @ExceptionHandler(gift.web.exception.notfound.NotFoundException.class)
+    public ResponseEntity<?> handleNotFoundException(NotFoundException e) {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
             ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, e.getMessage()));
     }
 
-    @ExceptionHandler(WishProductNotFoundException.class)
-    public ResponseEntity<?> handleWishProductNotFoundException(WishProductNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
-            ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, e.getMessage()));
-    }
-
-    @ExceptionHandler(MemberNotFoundException.class)
-    public ResponseEntity<?> handleMemberNotFoundException(MemberNotFoundException e) {
+    @ExceptionHandler(ForbiddenException.class)
+    public ResponseEntity<?> handleForbiddenException(ForbiddenException e) {
         return ResponseEntity.status(HttpStatus.FORBIDDEN).body(
             ProblemDetail.forStatusAndDetail(HttpStatus.FORBIDDEN, e.getMessage()));
-    }
-
-    @ExceptionHandler(CategoryNotFoundException.class)
-    public ResponseEntity<?> handleCategoryNotFoundException(CategoryNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
-            ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, e.getMessage()));
-    }
-
-    @ExceptionHandler(OptionNotFoundException.class)
-    public ResponseEntity<?> handleOptionNotFoundException(OptionNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
-            ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, e.getMessage()));
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
