+++ src/main/java/gift/web/exception/forbidden/ForbiddenException.java
@@ -1,0 +1,7 @@
+package gift.web.exception.forbidden;
+
+public class ForbiddenException extends RuntimeException {
+    public ForbiddenException(String message) {
+        super(message);
+    }
+}
