+++ src/main/java/gift/web/exception/notfound/NotFoundException.java
@@ -1,0 +1,7 @@
+package gift.web.exception.notfound;
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
