+++ src/main/java/gift/web/exception/unauthorized/UnauthorizedException.java
@@ -1,10 +1,10 @@
-package gift.web.exception;
+package gift.web.exception.unauthorized;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
 
 public class UnauthorizedException extends ResponseStatusException {
     public UnauthorizedException(String message) {
-        super(HttpStatus.UNAUTHORIZED, message);
+        super(HttpStatus.UNAUTHORIZED, "토큰이 유효하지 않습니다.");
     }
 }
